Adding -c switch to copy_new_paper.py

+++ copy_new_paper.py
@@ -13,7 +13,11 @@ import sys
 import shutil
 
 if len(sys.argv) < 2:
-    sys.exit("USAGE: python copy_new_paper.py [TARGET_DIR]")
+    sys.exit("USAGE: python3 copy_new_paper.py [TARGET_DIR]")
+
+create_clean = False
+if len(sys.argv) >= 3 and sys.argv[2] == "-c":
+    create_clean = True
 
 src_path = os.path.dirname(os.path.realpath(__file__))
 dest_path = sys.argv[1]
@@ -35,33 +39,36 @@ os.remove("copy_new_paper.py")
 os.remove("LICENSE")
 os.remove("README.md")
 
-# remove certain directories
+# remove .git directory
 shutil.rmtree(".git")
-shutil.rmtree("figures")
-shutil.rmtree("graphics")
-shutil.rmtree("table-data")
-shutil.rmtree("tables")
 
-# remove text in tex files in the sections directory
-os.chdir("sections")
+# create a "clean" version, removing unneeded files and content:
+if create_clean:
+    shutil.rmtree("figures")
+    shutil.rmtree("graphics")
+    shutil.rmtree("table-data")
+    shutil.rmtree("tables")
 
-keep_lines = (
-    "\\begin{abstract}",
-    "\\end{abstract}",
-    "\\section",
-    "\\subsection",
-    "\\label",
-)
-add_blank_lines = ("\\begin{abstract}", "\\label")
-for file in os.listdir("."):
-    if file.endswith(".tex"):
-        output = ""
-        with open(file, "r") as input:
-            for line in input:
-                if line.startswith(keep_lines):
-                    output += line
-                    if line.startswith(add_blank_lines):
-                        output += "\n"
-    tex_file = open(file, "w")
-    tex_file.write(output)
-    tex_file.close()
+    # remove text in tex files in the sections directory
+    os.chdir("sections")
+
+    keep_lines = (
+        "\\begin{abstract}",
+        "\\end{abstract}",
+        "\\section",
+        "\\subsection",
+        "\\label",
+    )
+    add_blank_lines = ("\\begin{abstract}", "\\label")
+    for file in os.listdir("."):
+        if file.endswith(".tex"):
+            output = ""
+            with open(file, "r") as input:
+                for line in input:
+                    if line.startswith(keep_lines):
+                        output += line
+                        if line.startswith(add_blank_lines):
+                            output += "\n"
+        tex_file = open(file, "w")
+        tex_file.write(output)
+        tex_file.close()
